services/backend/requirements.txt
aerich==0.5.8
aiomysql==0.0.21
aiosqlite==0.17.0
anyio==3.3.3
asgiref==3.4.1
asyncpg==0.24.0
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.70.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.2
iniconfig==1.1.1
iso8601==0.1.16
packaging==21.0
pluggy==1.0.0
py==1.10.0
pydantic==1.8.2
PyMySQL==0.9.3
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.5
pytz==2021.3
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
uvicorn==0.15.0
